+++ src/extension.ts
@@ -14,8 +14,8 @@ import * as nls from 'vscode-nls';
 import {CMakeCache} from '@cmt/cache';
 import {CMakeTools, ConfigureType, ConfigureTrigger} from '@cmt/cmake-tools';
 import {ConfigurationReader, TouchBarConfig} from '@cmt/config';
-import {CppConfigurationProvider} from '@cmt/cpptools';
-import {CMakeToolsFolderController, CMakeToolsFolder} from '@cmt/folders';
+import {CppConfigurationProvider, DiagnosticsCpptools} from '@cmt/cpptools';
+import {CMakeToolsFolderController, CMakeToolsFolder, DiagnosticsConfiguration, DiagnosticsSettings} from '@cmt/folders';
 import {
   Kit,
   USER_KITS_FILEPATH,
@@ -38,7 +38,8 @@ import {ProgressHandle, DummyDisposable, reportProgress} from '@cmt/util';
 import {DEFAULT_VARIANTS} from '@cmt/variant';
 import {expandString, KitContextVars} from '@cmt/expand';
 import paths from '@cmt/paths';
-import { CMakeDriver, CMakePreconditionProblems } from './drivers/driver';
+import {CMakeDriver, CMakePreconditionProblems} from './drivers/driver';
+import {platform} from 'os';
 
 nls.config({ messageFormat: nls.MessageFormat.bundle, bundleFormat: nls.BundleFormat.standalone })();
 const localize: nls.LocalizeFunc = nls.loadMessageBundle();
@@ -60,6 +61,15 @@ let _EXT_MANAGER: ExtensionManager|null = null;
 
 type CMakeToolsMapFn = (cmt: CMakeTools) => Thenable<any>;
 type CMakeToolsQueryMapFn = (cmt: CMakeTools) => Thenable<string | string[] | null>;
+
+interface Diagnostics {
+  os: string;
+  vscodeVersion: string;
+  cmtVersion: string;
+  configurations: DiagnosticsConfiguration[];
+  settings: DiagnosticsSettings[];
+  cpptoolsIntegration: DiagnosticsCpptools;
+}
 
 /**
  * A class to manage the extension.
@@ -117,6 +127,7 @@ class ExtensionManager implements vscode.Disposable {
         subs.push(new_cmt.onCodeModelChanged(FireLate, () => this._updateCodeModel(cmtFolder)));
         subs.push(new_cmt.onTargetNameChanged(FireLate, () => this._updateCodeModel(cmtFolder)));
         subs.push(new_cmt.onLaunchTargetNameChanged(FireLate, () => this._updateCodeModel(cmtFolder)));
+        subs.push(new_cmt.onActiveBuildPresetChanged(FireLate, () => this._updateCodeModel(cmtFolder)));
         this._codeModelUpdateSubs.set(new_cmt.folder.uri.fsPath, subs);
       }
       rollbar.takePromise('Post-folder-open', {folder: cmtFolder.folder}, this._postWorkspaceOpen(cmtFolder));
@@ -199,7 +210,8 @@ class ExtensionManager implements vscode.Disposable {
         this._codeModelUpdateSubs.set(cmtFolder.folder.uri.fsPath, [
           cmtFolder.cmakeTools.onCodeModelChanged(FireLate, () => this._updateCodeModel(cmtFolder)),
           cmtFolder.cmakeTools.onTargetNameChanged(FireLate, () => this._updateCodeModel(cmtFolder)),
-          cmtFolder.cmakeTools.onLaunchTargetNameChanged(FireLate, () => this._updateCodeModel(cmtFolder))
+          cmtFolder.cmakeTools.onLaunchTargetNameChanged(FireLate, () => this._updateCodeModel(cmtFolder)),
+          cmtFolder.cmakeTools.onActiveBuildPresetChanged(FireLate, () => this._updateCodeModel(cmtFolder))
         ]);
         rollbar.takePromise('Post-folder-open', {folder: cmtFolder.folder}, this._postWorkspaceOpen(cmtFolder));
       }
@@ -296,7 +308,7 @@ class ExtensionManager implements vscode.Disposable {
    */
   private readonly _configProvider = new CppConfigurationProvider();
   private _cppToolsAPI?: cpt.CppToolsApi;
-  private _configProviderRegister?: Promise<void>;
+  private _configProviderRegistered?: boolean = false;
 
   private _checkFolderArgs(folder?: vscode.WorkspaceFolder): CMakeToolsFolder | undefined {
     let cmtFolder: CMakeToolsFolder | undefined;
@@ -589,7 +601,7 @@ class ExtensionManager implements vscode.Disposable {
           await this.configureExtensionInternal(ConfigureTrigger.buttonNewKitsDefinition, cmt);
         } else {
           log.debug(localize('using.cache.to.configure.workspace.on.open', 'Using cache to configure workspace on open {0}', ws.uri.toString()));
-          await this.configureExtensionInternal(ConfigureTrigger.configureOnOpen, cmt);
+          await this.configureExtensionInternal(ConfigureTrigger.configureWithCache, cmt);
         }
       }
     }
@@ -724,15 +736,20 @@ class ExtensionManager implements vscode.Disposable {
           }
         }
 
+        const isMultiConfig = !!cache.get('CMAKE_CONFIGURATION_TYPES');
+        if (drv) {
+          drv.isMultiConfig = isMultiConfig;
+        }
+        const actualBuildType = isMultiConfig && cmt.useCMakePresets ? cmt.buildPreset?.configuration || null : cmt.activeVariant;
         const clCompilerPath = await findCLCompilerPath(env);
         this._configProvider.cpptoolsVersion = cpptools.getVersion();
         let codeModelContent;
         if (cmt.codeModelContent) {
           codeModelContent = cmt.codeModelContent;
-          this._configProvider.updateConfigurationData({ cache, codeModelContent, clCompilerPath, activeTarget: cmt.defaultBuildTarget, activeBuildTypeVariant: cmt.activeVariant, folder: cmt.folder.uri.fsPath });
+          this._configProvider.updateConfigurationData({ cache, codeModelContent, clCompilerPath, activeTarget: cmt.defaultBuildTarget, activeBuildTypeVariant: actualBuildType, folder: cmt.folder.uri.fsPath });
         } else if (drv && drv.codeModelContent) {
           codeModelContent = drv.codeModelContent;
-          this._configProvider.updateConfigurationData({ cache, codeModelContent, clCompilerPath, activeTarget: cmt.defaultBuildTarget, activeBuildTypeVariant: cmt.activeVariant, folder: cmt.folder.uri.fsPath });
+          this._configProvider.updateConfigurationData({ cache, codeModelContent, clCompilerPath, activeTarget: cmt.defaultBuildTarget, activeBuildTypeVariant: actualBuildType, folder: cmt.folder.uri.fsPath });
           this._projectOutlineProvider.updateCodeModel(
             cmt.workspaceContext.folder,
             codeModelContent,
@@ -742,14 +759,16 @@ class ExtensionManager implements vscode.Disposable {
             }
           );
         }
-        await this.ensureCppToolsProviderRegistered();
+        this.ensureCppToolsProviderRegistered();
         if (cpptools.notifyReady && this.cpptoolsNumFoldersReady < this._folders.size) {
           ++this.cpptoolsNumFoldersReady;
           if (this.cpptoolsNumFoldersReady === this._folders.size) {
             cpptools.notifyReady(this._configProvider);
+            this._configProvider.markAsReady();
           }
         } else {
           cpptools.didChangeCustomConfiguration(this._configProvider);
+          this._configProvider.markAsReady();
         }
       }
     });
@@ -1019,18 +1038,17 @@ class ExtensionManager implements vscode.Disposable {
     return this._folders.get(folder)?.useCMakePresets;
   }
 
-  async ensureCppToolsProviderRegistered() {
-    if (!this._configProviderRegister) {
-      this._configProviderRegister = this._doRegisterCppTools();
+  ensureCppToolsProviderRegistered() {
+    if (!this._configProviderRegistered) {
+      this._doRegisterCppTools();
+      this._configProviderRegistered = true;
     }
-    return this._configProviderRegister;
   }
 
-  async _doRegisterCppTools() {
-    if (!this._cppToolsAPI) {
-      return;
+  _doRegisterCppTools() {
+    if (this._cppToolsAPI) {
+      this._cppToolsAPI.registerCustomConfigurationProvider(this._configProvider);
     }
-    this._cppToolsAPI.registerCustomConfigurationProvider(this._configProvider);
   }
 
   private _cleanOutputChannel() {
@@ -1351,6 +1369,29 @@ class ExtensionManager implements vscode.Disposable {
     await logging.showLogFile();
   }
 
+  async logDiagnostics() {
+    telemetry.logEvent("logDiagnostics");
+    const configurations: DiagnosticsConfiguration[] = [];
+    const settings: DiagnosticsSettings[] = [];
+    for (const folder of this._folders.getAll()) {
+        configurations.push(await folder.getDiagnostics());
+        settings.push(await folder.getSettingsDiagnostics());
+    }
+
+    const result: Diagnostics = {
+      os: platform(),
+      vscodeVersion: vscode.version,
+      cmtVersion: util.thisExtensionPackage().version,
+      configurations,
+      cpptoolsIntegration: this._configProvider.getDiagnostics(),
+      settings
+    };
+    const output = logging.channelManager.get("CMake Diagnostics");
+    output.clear();
+    output.appendLine(JSON.stringify(result, null, 2));
+    output.show();
+  }
+
   activeFolderName(): string  {
     return this._folders.activeFolder?.folder.name || '';
   }
@@ -1617,6 +1658,7 @@ async function setup(context: vscode.ExtensionContext, progress?: ProgressHandle
     'setDefaultTarget',
     'resetState',
     'viewLog',
+    'logDiagnostics',
     'compileFile',
     'selectWorkspace',
     'tasksBuildCommand',