+++ src/util.ts
@@ -4,6 +4,7 @@ import * as fs from 'fs';
 import * as path from 'path';
 import * as vscode from 'vscode';
 import * as nls from 'vscode-nls';
+import { platform } from 'os';
 
 import {EnvironmentVariables, DebuggerEnvironmentVariable, execute} from '@cmt/proc';
 import rollbar from '@cmt/rollbar';
@@ -289,7 +290,7 @@ async function _killTree(pid: number) {
     }
     try {
       process.kill(pid, 'SIGINT');
-    } catch (e) {
+    } catch (e: any) {
       if (e.code === 'ESRCH') {
         // Do nothing. We're okay.
       } else {
@@ -313,8 +314,11 @@ export function splitCommandLine(cmd: string): string[] {
   return quoted_args!.map(arg => arg.replace(/\\(")/g, '$1').replace(/^"(.*)"$/g, '$1'));
 }
 
-export function isMultiConfGenerator(gen: string): boolean {
-  return gen.includes('Visual Studio') || gen.includes('Xcode');
+/**
+ * This is an initial check without atually configuring. It may or may not be accurate.
+ */
+export function isMultiConfGeneratorFast(gen: string): boolean {
+  return gen.includes('Visual Studio') || gen.includes('Xcode') || gen.includes('Multi-Config');
 }
 
 export class InvalidVersionString extends Error {}
@@ -648,9 +652,7 @@ export function isArrayOfString(x: any): x is string[] {
 }
 
 export function isNullOrUndefined(x?: any): boolean {
-  // Double equals provides the correct answer for 'null' and 'undefined'
-  // http://www.ecma-international.org/ecma-262/6.0/index.html#sec-abstract-equality-comparison
-  return x === null;
+  return (x === null || x === undefined);
 }
 
 export function isWorkspaceFolder(x?: any): boolean {
@@ -703,3 +705,30 @@ export function getRelativePath(file: string, dir: string): string {
   const joinedPath = "${workspaceFolder}/".concat(relPathDir);
   return joinedPath;
 }
+
+async function getHostSystemName(): Promise<string> {
+  if (platform() === "win32") {
+    return "Windows";
+  } else {
+    const result = await execute('uname', ['-s']).result;
+    if (result.retc === 0) {
+      return result.stdout.trim();
+    } else {
+      return 'unknown';
+    }
+  }
+}
+
+function memoize<T>(fn: () => T) {
+  let result: T;
+
+  return () => {
+    if (result) {
+      return result;
+    } else {
+      return result = fn();
+    }
+  };
+}
+
+export const getHostSystemNameMemo = memoize(getHostSystemName);